Teljesen új téma importálása, főoldal törzsének megvalósítása, CRUD táblázat a jelenlegi pilótákkal létrehozva, MYSQL adatbázishoz előkészítve

+++ database/migrations/2025_10_29_203139_create_results_table.php
@@ -6,29 +6,23 @@ use Illuminate\Support\Facades\Schema;
 
 return new class extends Migration
 {
-    /**
-     * Run the migrations.
-     */
     public function up(): void
     {
         Schema::create('results', function (Blueprint $table) {
             $table->id();
-            $table->date('race_date'); // datum mező
-            $table->integer('pilot_id'); // pilotaaz mező
-            $table->integer('position')->nullable(); // helyezes mező
-            $table->string('issue')->nullable(); // hiba mező
-            $table->string('team'); // csapat mező
-            $table->string('car_type'); // tipus mező
-            $table->string('engine'); // motor mező
+            $table->date('race_date');
+            $table->unsignedBigInteger('pilot_id');
+            $table->integer('position')->nullable();
+            $table->string('issue')->nullable();
+            $table->string('team');
+            $table->string('car_type');
+            $table->string('engine');
             $table->timestamps();
             
             $table->foreign('pilot_id')->references('pilot_id')->on('pilots');
         });
     }
 
-    /**
-     * Reverse the migrations.
-     */
     public function down(): void
     {
         Schema::dropIfExists('results');
